[mcp] Move to /tools
Moves to a tools directory.
